+++ src/Request/ODataRequestDataInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace SnelstartPHP\Request;
+
+interface ODataRequestDataInterface
+{
+    public function getHttpCompatibleQueryString(): string;
+}